Reorganize annotated_arg/annotated_ptr
