+++ sycl/include/sycl/ext/oneapi/common_annotated_properties/common_properties.hpp
@@ -20,9 +20,6 @@ namespace experimental {
 template <typename T, typename PropertyListT> class annotated_arg;
 template <typename T, typename PropertyListT> class annotated_ptr;
 
-//===----------------------------------------------------------------------===//
-//        Common properties of annotated_arg/annotated_ptr
-//===----------------------------------------------------------------------===//
 struct register_map_key {
   using value_t = property_value<register_map_key>;
 };
@@ -338,36 +335,6 @@ struct check_property_list<T, Prop, Props...>
                 "Property is invalid for the given type.");
 };
 
-//===----------------------------------------------------------------------===//
-//        Specific properties of annotated_ptr
-//===----------------------------------------------------------------------===//
-struct alignment_key {
-  template <int K>
-  using value_t = property_value<alignment_key, std::integral_constant<int, K>>;
-};
-
-template <int K> inline constexpr alignment_key::value_t<K> alignment;
-
-template <> struct is_property_key<alignment_key> : std::true_type {};
-
-template <typename T, typename PropertyListT>
-struct is_property_key_of<alignment_key, annotated_ptr<T, PropertyListT>>
-    : std::true_type {};
-
-namespace detail {
-
-template <> struct PropertyToKind<alignment_key> {
-  static constexpr PropKind Kind = PropKind::Alignment;
-};
-
-template <> struct IsCompileTimeProperty<alignment_key> : std::true_type {};
-
-template <int N> struct PropertyMetaInfo<alignment_key::value_t<N>> {
-  static constexpr const char *name = "sycl-alignment";
-  static constexpr int value = N;
-};
-
-} // namespace detail
 
 //===----------------------------------------------------------------------===//
 //   Utility type trait for annotated_arg/annotated_ptr deduction guide
