+++ sycl/include/sycl/sycl.hpp
@@ -65,9 +65,10 @@
 #include <sycl/ext/intel/experimental/pipe_properties.hpp>
 #include <sycl/ext/intel/experimental/pipes.hpp>
 #include <sycl/ext/intel/usm_pointers.hpp>
+#include <sycl/ext/oneapi/common_annotated_properties/common_properties.hpp>
 #include <sycl/ext/oneapi/annotated_arg/annotated_arg.hpp>
-#include <sycl/ext/oneapi/annotated_arg/annotated_ptr.hpp>
-#include <sycl/ext/oneapi/annotated_arg/properties.hpp>
+#include <sycl/ext/oneapi/annotated_ptr/annotated_ptr.hpp>
+#include <sycl/ext/oneapi/annotated_ptr/annotated_ptr_specific_properties.hpp>
 #include <sycl/ext/oneapi/device_global/device_global.hpp>
 #include <sycl/ext/oneapi/device_global/properties.hpp>
 #include <sycl/ext/oneapi/experimental/ballot_group.hpp>
