+++ sycl/test/extensions/annotated_properties/annotated_properties.cpp
