+++ sycl/test/extensions/annotated_ptr/annotated_ptr.cpp
